2D Sweep Updates For Outputs And Inputs
Sequence Helper Functions reworked to include looping; acquisitions longer than 65 us will now be split into uniform sections and looped to allow for longer acquisitions of the same voltage value. Similarly, setting the offset for the AWG with the QCM has been reworked in this fashion, allowing for voltage values to be played longer than 65 us. In general, not only do these updates generalize the code by allowing for longer pulses and acquisitions, the addition of looping reduces the number of commands sent to qblox for the same amount of steps in a given 2D sweep, allowing for 2D sweeps to contain more points in general.